specs-go/v1/config: pointer-ify optional "Created" field
PR #534 changed Created from string to time.Time field.
This neutered the `omitempty` property on it, as a missing
field is now populated as the default time.Time value.

This commit turns the field into a pointer, in order to keep
the empty-value semantic for `omitempty`.

Fixes https://github.com/opencontainers/image-spec/issues/626

Signed-off-by: Luca Bruno <luca.bruno@coreos.com>