image-layout.md, annotations.md: restrict scope of ref.name
By restricting the scope of the annotation
`org.opencontainers.image.ref.name` to only image indexes appearing in
the context of an image layout, we can greatly simplify processing and
selection of target content. While other properties may apply down, we
ensure that naming is restricted to a single "entity", the `index.json`,
ambiguity for a given reference is reduced.

This does not apply to other annotations, which may "apply downward" to
other resources referenced. The application of annotation downward is
specific to a particular annotation.

Signed-off-by: Stephen J Day <stephen.day@docker.com>