spec: Clarify that DNS-naming and signing have not been specified
We don't have these yet.  I'd rather remove them, because:

* We already track roadmap issues in the README [1] (which is not part
  of the spec itself).
* We also have issue-labels for these features [2,3].
* We have entries in the scope table for these features [4].
* We can recover the current wording from version control if/when we
  land spec sections to back them.

But at least three maintainers are in favor of keeping a note of some
sort [5,6,7], and the only direct response to my desire to move the
references to the existing roadmap section was:

On Thu, Apr 20, 2017 at 04:09:32AM -0700, Jonathan Boulle wrote [8]:
> Since the roadmap to 1.0 doesn't reference these, please just add
> work-not-yet-started here in this list.

I still don't see why that section needs to be 1.0-specific,
especially since the only accurate part of the current roadmap body is
the link to milestones which are not 1.0-specific.  But I've already
made the arguments I can think of, and for whatever reason the
maintainers aren't convinced.

[1]: https://github.com/opencontainers/image-spec/blob/v1.0.0-rc5/README.md#roadmap
[2]: https://github.com/opencontainers/image-spec/labels/component%2Fnaming%20spec
[3]: https://github.com/opencontainers/image-spec/labels/component%2Fsigning%20spec
[4]: https://www.opencontainers.org/about/oci-scope-table
[5]: https://github.com/opencontainers/image-spec/pull/643#issuecomment-292922990
[6]: https://github.com/opencontainers/image-spec/pull/643#issuecomment-292930686
[7]: https://github.com/opencontainers/image-spec/pull/643#issuecomment-295385422
[8]: https://github.com/opencontainers/image-spec/pull/643#issuecomment-295687098

Signed-off-by: W. Trevor King <wking@tremily.us>