image-index: remove platform.features field
While the intended use of the `platform.features` field was to express
the availability of hardware instructions on CPUs, the practicality of
expressing and implementing such a standard has been called into
question. Much of the functionality is usually hidden deep within kernel
code or is defined by hardware vendors in ways that will make the field
unusable without future additions. Applications that can make use of such
field will likely have other requirements on top of this field that are
specific to a deployment.

In that regard, it is best to remove this field and defer to hardware to
make the decision about what can be run.

Signed-off-by: Stephen J Day <stephen.day@docker.com>