annotations: Use SPDX License Expressions for licenses
Instead of comma-separated short identifiers, which have unclear
semantics (are the delimiters AND or OR?).  I don't see any discussion
of the syntax for this field in [1] (which landed it), but I'd floaded
license expressions before in the sub-thread starting at [2].  Greg
had pushed back against my earlier proposal (licensing information on
descriptors) with [3]:

> No, that's not going to work at all, you can't properly describe the
> license for a whole layer in any form of a string that could be
> standardized or parsed. SPDX is great for describing the individual
> licenses of things, but not for a collection of things, which almost
> always has an arbitrary license (example, what's the license, in a
> simple string, for a Ubuntu base layer?)

But SPDX License Expression are both more expressive and better
defined than the current comma delimiters.  Everything you could have
said with the comma-delimited string you can say more clearly with a
SPDX License Expression.  And because the syntax is not OCI-specific,
you're more likely to be able to find tooling that handles these
values out of the box.

[1]: https://github.com/opencontainers/image-spec/pull/636
[2]: https://github.com/opencontainers/image-spec/pull/501#discussion_r96761939
[3]: https://github.com/opencontainers/image-spec/pull/501#discussion_r97056075

Signed-off-by: W. Trevor King <wking@tremily.us>