Add artifactType to image index
It would be useful for implementations which generate multiple artifacts, each
with their own platform or other metadata and perhaps, associated signatures,
attestations, to be able to index them all in a single image index.

As an implementer, I'm considering having a top level "package" stored as an
image index, which then references all the artifacts in the package as artifacts
(image manifests). The artifactType of the image index would be
"application/...package", and the artifactType of the image manifests could be:

* "application/...plugin" with annotations and/or platform fields set to
  indicate that it is the plugin for, e.g., macOS with Apple Silicon
  (darwin-arm64)
* "application/...schema" with annotations indicating this is a separate artifact
* ... other artifact types as needed

Using an image index in this use case is helpful as we would want to associate
SBOMs, signatures to the individual artifacts references, so that a client only
needs to download and verify the artifacts it consumes selects from the index.

Signed-off-by: Aaron Friel <mayreply@aaronfriel.com>