Reduce maintenance burden by letting setup-go track our Go versions
`actions/setup-go` supports "stable" and "oldstable" as automatic aliases for the latest release and the previous to latest release, which was our intent with two of these three hard-coded versions.  The third is hard-coded to match our `go.mod`, which `actions/setup-go` *also* supports reading from, so we can be completely DRY here.

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>