Docs: Update pandoc for rendering documentation
The main reason for this is to fix the links in the output docs, which currently all 404.
This upgrades pandoc to 3.1, which isn't the latest, due to a regression in the relative links between markdown files.
I pulled out the "header.html" logic, injected the title directly, and configured the build to embed images into the output doc for better portability (and the ability to generate dev docs).

Signed-off-by: Brandon Mitchell <git@bmitch.net>